requirements.txt
langchain
langchain-community
python-dotenv
pytest
langchain-ollama